% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamUtils.R
\name{bamflag}
\alias{bamflag}
\title{Returns matrix of bits from BAM flags}
\usage{
bamflag(reads)
}
\arguments{
\item{reads}{GenomicRanges holding the reads}
}
\value{
matrix of bits from BAM flags
}
\description{
Shortcut function: assumes reads are GappedAlignments with flag variable or actual integers representing BAM flag
}
