% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamUtils-package.R
\docType{package}
\name{bamUtils-package}
\alias{bamUtils}
\alias{bamUtils-package}
\title{bamUtils: Utility functions for manipulating BAMs}
\description{
Utility functions for manipulating BAMs
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{http://github.com/mskilab/bamUtils/issues}
}

}
\author{
\strong{Maintainer}: Marcin Imielinski \email{mai9037@med.cornell.edu}

Authors:
\itemize{
  \item Jeremiah Wala \email{jwala@broadinstitute.org}
}

}
