% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamUtils.R
\name{chunk}
\alias{chunk}
\title{chunk}
\usage{
chunk(from, to = NULL, by = 1, length.out = NULL)
}
\arguments{
\item{from}{integer Where to begin sequence}

\item{to}{integer To end sequence (default = NULL)}

\item{by}{integer Interval to space sequence (default = 1)}

\item{length.out}{integer Number of desired chunks, i.e. nrows of output matrix (default = NULL)}
}
\value{
2-column matrix of indices, each row representing a chunk
}
\description{
Internal function takes same input as seq (from, to, by, length.out) and outputs a 2 column matrix of indices
corresponding to "chunks"
}
\author{
Marcin Imielinski
}
