% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamUtils.R
\name{count.clips}
\alias{count.clips}
\title{Return data.frame with fields of "right" soft clips and "left" soft clips}
\usage{
count.clips(reads, hard = FALSE)
}
\arguments{
\item{reads}{GenomicRanges holding the reads}

\item{hard}{boolean Option counts hard clips (default == FALSE)}
}
\description{
Takes GRanges or GappedAlignments object and uses cigar field (or takes character vector of cigar strings)
and returns data.frame with fields (for character input)
$right.clips number of "right" soft clips (e.g. cigar 89M12S)
$left.clips number of "left" soft clips (e.g. cigar 12S89M), 
or appends these fields to the reads object
}
