% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamUtils.R
\name{get.mate.gr}
\alias{get.mate.gr}
\title{returns GRanges corresponding to mates of reads}
\usage{
get.mate.gr(reads)
}
\arguments{
\item{reads}{GRanges or data.table/data.frame Input reads}
}
\value{
\code{GRanges} corresponding to mates of reads
}
\description{
Inputs GRanges or data.frame/data.table of reads. Outputs GRanges corresponding to mates of reads.
}
