% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamUtils.R
\name{get.pairs.grl}
\alias{get.pairs.grl}
\title{Get coverage as GRanges from BAM on custom set of GRanges}
\usage{
get.pairs.grl(reads, pairs.grl.split = TRUE, verbose = FALSE)
}
\arguments{
\item{reads}{GRanges holding reads}

\item{pairs.grl.split}{boolean Flag to return as GRangesList if TRUE (default == TRUE)}

\item{verbose}{boolean verbose flag (default == FALSE)}
}
\description{
Takes reads object and returns GRangesList with each read and its mate (if exists)
}
