% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamUtils.R
\name{oneoffs}
\alias{oneoffs}
\title{Calls samtools mpileup to dump tsv of "one off" variants / sites}
\usage{
oneoffs(out.file, bam, ref, min.bq = 30, min.mq = 60, indel = FALSE,
  chunksize = 10000, verbose = TRUE)
}
\arguments{
\item{out.file}{file to dump tsv to}

\item{bam}{bam file path}

\item{ref}{fasta path}

\item{min.bq}{integer minimum base quality}

\item{min.mq}{integer minimum mapping quality}

\item{indel}{logical flag whether to collect one off indels (default is substitution)}

\item{chunksize}{number of mpileup lines to put into memory}

\item{verbose}{logical flag}
}
\description{
Calls samtools mpileup to dump tsv of "one off" variants / sites (i.e. that are present in exactly one read per site)
}
\note{
The denominator (ie total reads) is just the sum of counts$records
}
