% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamUtils.R
\name{read.bam}
\alias{read.bam}
\title{read.bam}
\usage{
read.bam(bam, intervals = NULL, gr = intervals, all = FALSE, bai = NULL,
  pairs.grl = TRUE, stripstrand = TRUE, what = scanBamWhat(),
  unpack.flag = FALSE, verbose = FALSE, tag = NULL, isPaired = NA,
  isProperPair = NA, isUnmappedQuery = NA, hasUnmappedMate = NA,
  isNotPassingQualityControls = NA, isDuplicate = FALSE,
  isValidVendorRead = TRUE, pairs.grl.split = TRUE, as.data.table = FALSE,
  ignore.indels = FALSE, size.limit = 1e+06, ...)
}
\arguments{
\item{bam}{Input bam file. Advisable to make "bam" a BamFile instance instead of a plain string, so that the index does not have to be reloaded.}

\item{intervals}{GRanges of intervals to retrieve}

\item{gr}{GRanges of intervals to retrieve}

\item{stripstrand}{Flag to ignore strand information on the query intervals. Default TRUE}

\item{what}{What fields to pull down from BAM. Default \code{scanBamWhat()}}

\item{unpack.flag}{Add features corresponding to read flags. Default FALSE}

\item{verbose}{Increase verbosity}

\item{tag}{Additional tags to pull down from the BAM (e.g. 'R2')}

\item{isPaired}{See documentation for \code{scanBamFlag}. Default NA}

\item{isProperPair}{See documentation for \code{scanBamFlag}. Default NA}

\item{isUnmappedQuery}{See documentation for \code{scanBamFlag}. Default NA}

\item{hasUnmappedMate}{See documentation for \code{scanBamFlag}. Default NA}

\item{isNotPassingQualityControls}{See documentation for \code{scanBamFlag}. Default NA}

\item{isDuplicate}{See documentation for \code{scanBamFlag}. Default FALSE}

\item{isValidVendorRead}{See documentation for \code{scanBamFlag}. (default = TRUE)}

\item{pairs.grl.split}{boolean returns as GRangesList if TRUE (default = TRUE)}

\item{as.data.table}{Return reads in the form of a data.table rather than GRanges/GRangesList}

\item{ignore.indels}{messes with cigar to read BAM with indels removed. Useful for breakpoint mapping on contigs}

\item{...}{passed to \code{scanBamFlag}}

\item{bami}{Input bam index file.}
}
\value{
Reads in one of GRanges, GRangesList or data.table
}
\description{
Read BAM file into GRanges or data.table

Wrapper around Rsamtools bam scanning functions,
by default, returns GRangesList of read pairs for which <at least one> read lies in the supplied interval
}
