% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamUtils.R
\name{read_vcf}
\alias{read_vcf}
\title{parses VCF into GRanges or data.table}
\usage{
read_vcf(fn, gr = NULL, hg = "hg19", geno = NULL, swap.header = NULL,
  verbose = FALSE, add.path = FALSE, tmp.dir = "~/temp/.tmpvcf", ...)
}
\arguments{
\item{fn}{info}

\item{gr}{GRanges}

\item{hg}{string Human reference genome (default = 'hg19')}

\item{geno}{(default = NULL)  ### 'geno()' method from 'VCF-class'; ‘geno(x, withDimnames=TRUE)’, ‘geno(x) <- value’: Gets a SimpleList of genotype data.  ‘value’ is a SimpleList.}

\item{swap.header}{(default = NULL)}

\item{verbose}{boolean Flag (default = FALSE)}

\item{add.path}{boolean Flag (default = FALSE)}

\item{tmp.dir}{string Path (defautl = '~/temp/.tmpvcf')}
}
\description{
Wrapper around Bioconductor VariantAnnotation. Reads VCF into GRanges or data.table format
}
\author{
Marcin Imielinski
}
