% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamUtils.R
\name{varbase}
\alias{varbase}
\title{Returns variant bases and ranges from GRanges or GappedAlignments input}
\usage{
varbase(reads, soft = TRUE, verbose = TRUE)
}
\arguments{
\item{reads}{GenomicRanges to extract variants from}

\item{soft}{boolean Flag to include soft-clipped matches (default == TRUE)}

\item{verbose}{boolean verbose flag (default == TRUE)}
}
\description{
Takes GRanges or GappedAlignments object "reads" and uses cigar, MD, seq fields
to return variant bases and ranges

Teturns GRangesList (of same length as input) of variant base positions with character vector 
$varbase field populated with variant bases for each GRanges item in grl[[k]], 
with the following handling for insertions, deletions, and substitution GRange's:

Substitutions: nchar(gr$varbase) = width(gr) of the corresponding var
Insertions: nchar(gr$varbase)>=1, width(gr) ==0
Deletions: gr$varbase = '', width(gr)>=1

Each GRanges also has $type flag which shows the cigar string code for the event i.e.
S = soft clip --> varbase represents clipped bases
I = insertion --> varbase represents inserted bases
D = deletion --> varbase is empty
X = mismatch --> varbase represents mismatched bases
}
