% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamUtils.R
\name{bam.cov.gr}
\alias{bam.cov.gr}
\title{Get coverage as GRanges from BAM on custom set of GRanges}
\usage{
bam.cov.gr(bam, gr, bai = NULL, count.all = FALSE, isPaired = TRUE,
  isProperPair = TRUE, isUnmappedQuery = FALSE, hasUnmappedMate = FALSE,
  isNotPassingQualityControls = FALSE, isDuplicate = FALSE, mc.cores = 1,
  chunksize = 10, verbose = FALSE, ...)
}
\arguments{
\item{bam}{string Input BAM file. Advisable to make the input BAM a BamFile instance instead of a plain string, so that the index does not have to be reloaded.}

\item{gr}{GRanges of intervals to retrieve}

\item{bai}{string Input BAM index file.}

\item{isPaired}{boolean Flag indicates whether unpaired (FALSE), paired (TRUE), or any (NA) read should be returned. See documentation for Rsamtools::scanBamFlag(). (default == NA)}

\item{isProperPair}{boolean Flag indicates whether improperly paired (FALSE), properly paired (TRUE), or any (NA) read should be returned. A properly paired read is defined by the alignment algorithm and might, e.g., represent reads aligning to identical reference sequences and with a specified distance. See documentation for Rsamtools::scanBamFlag(). (default == NA)}

\item{isUnmappedQuery}{boolean Flag indicates whether unmapped (TRUE), mapped (FALSE), or any (NA) read should be returned. See documentation for Rsamtools::scanBamFlag(). (default == NA)}

\item{hasUnmappedMate}{boolean Flag indicates whether reads with mapped (FALSE), unmapped (TRUE), or any (NA) mate should be returned. See documentation for Rsamtools::scanBamFlag(). (default == NA)}

\item{isNotPassingQualityControls}{boolean Flag indicates whether reads passing quality controls (FALSE), reads not passing quality controls (TRUE), or any (NA) read should be returned. See documentation for Rsamtools::scanBamFlag(). (default == NA)}

\item{isDuplicate}{boolean Flag indicates that un-duplicated (FALSE), duplicated (TRUE), or any (NA) reads should be returned. 'Duplicated' reads may represent PCR or optical duplicates. See documentation for Rsamtools::scanBamFlag(). (default == FALSE)}

\item{mc.cores}{integer Number of cores in \code{mclapply} call}

\item{chunksize}{integer How many intervals to process per core (default == 10)}

\item{verbose}{boolean verbose flag (default == FALSE)}

\item{...}{futher arguments passed to Rsamtools::scanBamFlag()}
}
\value{
GRanges parallel to input GRanges, but with metadata filled in.
}
\description{
Gets coverage from BAM in supplied GRanges using 'countBam()', returning GRanges with coverage counts in
each of the provided GRanges (different from 'bamUtils::bam.cov()') specified as the 
columns $file, $records, and $nucleotides in the values field

Basically a wrapper for 'Rsamtools::countBam()' with some standard settings for 'Rsamtools::ScanBamParams()'
}
