% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamUtils.R
\name{bam.cov.tile}
\alias{bam.cov.tile}
\title{Get coverage as GRanges from BAM on genome tiles across seqlengths of genome}
\usage{
bam.cov.tile(
  bam.file,
  window = 100,
  chunksize = 1e+05,
  min.mapq = 30,
  verbose = TRUE,
  max.tlen = 10000,
  st.flag = "-f 0x02 -F 0x10",
  fragments = TRUE,
  do.gc = FALSE,
  midpoint = TRUE,
  bai = NULL
)
}
\arguments{
\item{bam.file}{string Input BAM file}

\item{window}{integer Window size (in bp) (default = 1e2)}

\item{chunksize}{integer Size of window (default = 1e5)}

\item{min.mapq}{integer Minimim map quality reads to consider for counts (default = 30)}

\item{verbose}{boolean Flag to increase vebosity (default = TRUE)}

\item{max.tlen}{integer Maximum paired-read insert size to consider (default = 1e4)}

\item{st.flag}{string Samtools flag to filter reads on (default = '-f 0x02 -F 0x10')}

\item{fragments}{boolean flag (default = FALSE) detremining whether to compute fragment (i.e. proper pair footprint aka insert) density or read density}

\item{do.gc}{boolean Flag to execute garbage collection via 'gc()' (default = FALSE)}

\item{midpoint}{boolean Flag if TRUE will only use the fragment midpoint, if FALSE will count all bins that overlap the fragment (default = TRUE)}
}
\value{
GRanges of "window" bp tiles across seqlengths of bam.file with meta data field $counts specifying fragment counts centered (default = TRUE)
in the given bin.
}
\description{
Quick way to get tiled coverage via piping to samtools (e.g. ~10 CPU-hours for 100bp tiles, 5e8 read pairs)

Gets coverage for window size "window" [bp], pulling "chunksize" records at a time and incrementing bin
corresponding to midpoint or overlaps of corresponding (proper pair) fragment (uses TLEN and POS for positive strand reads that are part of a proper pair)
}
