% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamUtils.R
\name{is.paired.end}
\alias{is.paired.end}
\title{Check if BAM file is paired end by using 0x1 flag}
\usage{
is.paired.end(bams)
}
\arguments{
\item{bams}{vector of input BAMs}
}
\value{
boolean returns TRUE if BAM file is paired-end, returns FALSE if BAM not paired-end
}
\description{
Check if BAM file is paired-end by using 0x1 flag,
pipes to 'samtools' via command line
}
