% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamUtils.R
\name{read_vcf}
\alias{read_vcf}
\title{parses VCF into GRanges or data.table}
\usage{
read_vcf(fn, gr = NULL, hg = "hg19", geno = NULL, swap.header = NULL,
  verbose = FALSE, add.path = FALSE, tmp.dir = "~/temp/.tmpvcf", ...)
}
\arguments{
\item{fn}{argument to parse via bcftools}

\item{gr}{GRanges input GRanges (default = NULL)}

\item{hg}{string Human reference genome (default = 'hg19')}

\item{geno}{boolean Flag whether to pull the genotype information information in the GENO vcf fields (default = NULL)}

\item{swap.header}{string Pathn to another VCF file (in case of VCF with malformed header)(default = NULL)}

\item{verbose}{boolean Flag (default = FALSE)}

\item{add.path}{boolean Flag to add the path of the current VCF file to the output (default = FALSE)}

\item{tmp.dir}{string Path to directory for temporary files (default = '~/temp/.tmpvcf')}

\item{...}{extra parameters}
}
\description{
Wrapper around Bioconductor VariantAnnotation. Reads VCF into GRanges or data.table format
}
\author{
Marcin Imielinski
}
