% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamUtils.R
\name{varcount}
\alias{varcount}
\title{Wrapper around applyPileups}
\usage{
varcount(bams, gr, min.mapq = 0, min.baseq = 20, max.depth = 500,
  indel = FALSE, ...)
}
\arguments{
\item{bams}{character vector of paths to bam files}

\item{gr}{GRanges of (width=1) sites i.e. intervals at which to compute base coujnts}

\item{min.mapq}{integer Minimal mapping quality at which to compute bases}

\item{max.depth}{integer Maximum read depth to consider}

\item{indel}{boolean Flag whether to consider indels (default FALSE)}

\item{...}{other args be passed to read.bam()}

\item{max.baseq}{integer Minimal base qualitya t which to compute bases}
}
\value{
GRanges annotated with fields $alt.count.t, $ref.count.t, $alt.count.n, $ref.count.n
}
\description{
Takes in vector of bam paths or GRanges corresponding to sites / territories to query,
and outputs a list with fields:

$counts = 3D matrix of base counts
(A, C, G, T, N) x sites x bams subject to mapq and baseq thresholds



varcount() relies upon varbase()
}
\author{
Marcin Imielinski
}
